refactor : add correct importation in e2e tests
